implement UI component for sidebar
